bugfix: [SEO/Shows] Fix 404 link back to inactive fairs
